# store-app/data/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Store App API",
    "description": "APIهای عمومی فروشگاه — نسخه Q4 2026",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Same origin" }
  ],
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": ["catalog"],
        "summary": "لیست محصولات",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "لیست محصولات" }
        }
      }
    },
...
  }
}